decompress.c: change uint to size_t
With size_t, gcc produces better code.  I tried to compare the assembly
output, but register allocation changes, so the diff is not insightful.
At least I can see one 3-arg LEA instruction.  The thing is, gcc may
assume that mlen+=3 may overflow, so with uint it can't be transformed
into lea 3(out,mlen),out.

-relz4 0.1                  27 MB/s  3640 MB/s    78126076  36.86 12 files
+relz4 0.1                  28 MB/s  3741 MB/s    78126076  36.86 12 files
 relz4 0.1                  27 MB/s  3664 MB/s    78126076  36.86 12 files (asm)

Can't beat gcc again! :(

+++ decompress.c
@@ -4,11 +4,11 @@
 static inline uchar *decompress(const uchar *src, size_t srcSize, uchar *out)
 {
     const uchar *srcLast5 = src + srcSize - 5;
-    uint tok = *src;
-    uint llen = tok >> 4;
+    size_t tok = *src;
+    size_t llen = tok >> 4;
     src += 2;
     while (1) {
-	uint mlen = tok;
+	size_t mlen = tok;
 	tok = src[-1];
 	if (likely(mlen < 0xf0)) {
 	    memcpy(out, src, 16);
@@ -26,7 +26,7 @@ static inline uchar *decompress(const uchar *src, size_t srcSize, uchar *out)
 	}
 	if (unlikely(src > srcLast5))
 	    break;
-	uint moff = load16le(src);
+	size_t moff = load16le(src);
 	llen = tok >> 4;
 	const uchar *ref = out - moff;
 	memcpy(out + 0, ref - 16, 16);
